Run tests for the new Stop(bool) overload only on .NETCoreApp
