+++ src/libraries/System.ServiceProcess.ServiceController/tests/ServiceControllerTests.cs
@@ -7,7 +7,7 @@ namespace System.ServiceProcess.Tests
 {
     [OuterLoop(/* Modifies machine state */)]
     [SkipOnTargetFramework(TargetFrameworkMonikers.NetFramework, "Persistent issues starting test service on NETFX")]
-    public class ServiceControllerTests : IDisposable
+    public partial class ServiceControllerTests : IDisposable
     {
         private const int connectionTimeout = 30000;
         private readonly TestServiceProvider _testService;
@@ -76,27 +76,6 @@ namespace System.ServiceProcess.Tests
         {
             var controller = new ServiceController(_testService.TestServiceName);
             Assert.Throws<ArgumentNullException>(() => controller.Start(new string[] { null } ));
-        }
-
-        [ConditionalFact(nameof(IsProcessElevated))]
-        public void Stop_FalseArg_WithDependentServices_ThrowsInvalidOperationException()
-        {
-            var controller = new ServiceController(_testService.TestServiceName);
-            controller.WaitForStatus(ServiceControllerStatus.Running, _testService.ControlTimeout);
-            Assert.Throws<InvalidOperationException>(() => controller.Stop(stopDependentServices: false));
-        }
-
-        [ConditionalFact(nameof(IsProcessElevated))]
-        public void Stop_TrueArg_WithDependentServices_StopsTheServiceAndItsDependents()
-        {
-            var controller = new ServiceController(_testService.TestServiceName);
-            controller.WaitForStatus(ServiceControllerStatus.Running, _testService.ControlTimeout);
-
-            controller.Stop(stopDependentServices: true);
-            controller.WaitForStatus(ServiceControllerStatus.Stopped, _testService.ControlTimeout);
-
-            Assert.Equal(ServiceControllerStatus.Stopped, controller.Status);
-            Assert.All(controller.DependentServices, service => Assert.Equal(ServiceControllerStatus.Stopped, service.Status));
         }
 
         [ConditionalFact(nameof(IsProcessElevated))]
